Moved the estimates by job role job, utils, models, tests and validation to new folder structure.
