+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/tests/utils/test_estimate_ind_cqc_filled_posts_by_job_role_models_primary_service_rolling_sum.py
@@ -1,14 +1,17 @@
 import unittest
 
-from utils import utils
-from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
-from utils.estimate_filled_posts_by_job_role_utils.models import (
+from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.utils.models import (
     primary_service_rolling_sum as job,
 )
-from tests.test_file_data import EstimateJobRolesPrimaryServiceRollingSumData as Data
-from tests.test_file_schemas import (
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_data import (
+    EstimateJobRolesPrimaryServiceRollingSumData as Data,
+)
+from projects._03_independent_cqc.unittest_data.ind_cqc_test_file_schemas import (
     EstimateJobRolesPrimaryServiceRollingSumSchemas as Schemas,
 )
+
+from utils import utils
+from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
 
 
 class EstimateJobRolesPrimaryServiceRollingSumTests(unittest.TestCase):
