+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/utils/models/interpolation.py
@@ -2,7 +2,7 @@ from pyspark.sql import DataFrame, functions as F
 from typing import List
 
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
-from utils.estimate_filled_posts_by_job_role_utils.utils import (
+from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.utils.utils import (
     unpack_mapped_column,
     create_map_column,
     pivot_job_role_column,
