+++ projects/_03_independent_cqc/_07_estimate_filled_posts_by_job_role/utils/models/primary_service_rolling_sum.py
@@ -1,7 +1,7 @@
 from pyspark.sql import DataFrame, functions as F, Window
 
 from utils.column_names.ind_cqc_pipeline_columns import IndCqcColumns as IndCQC
-from utils.estimate_filled_posts_by_job_role_utils.utils import (
+from projects._03_independent_cqc._07_estimate_filled_posts_by_job_role.utils.utils import (
     create_map_column,
     pivot_job_role_column,
 )
