+++ projects/_03_independent_cqc/unittest_data/ind_cqc_test_file_schemas.py
@@ -5,8 +5,11 @@ from pyspark.sql.types import (
     StructType,
     StructField,
     StringType,
+    MapType,
+    LongType,
     IntegerType,
     FloatType,
+    DoubleType,
     DateType,
     ArrayType,
 )
@@ -24,6 +27,7 @@ from utils.column_names.ind_cqc_pipeline_columns import (
     ArchivePartitionKeys as ArchiveKeys,
     IndCqcColumns as IndCQC,
 )
+from utils.column_values.categorical_column_values import MainJobRoleLabels
 
 
 @dataclass
@@ -431,3 +435,681 @@ class ArchiveFilledPostsEstimates:
             StructField(ArchiveKeys.archive_timestamp, StringType(), True),
         ]
     )
+
+
+@dataclass
+class EstimateIndCQCFilledPostsByJobRoleSchemas:
+    estimated_ind_cqc_filled_posts_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+            StructField(IndCQC.primary_service_type, StringType(), True),
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(IndCQC.ascwds_workplace_import_date, DateType(), True),
+            StructField(IndCQC.estimate_filled_posts, DoubleType(), True),
+            StructField(
+                IndCQC.registered_manager_names, ArrayType(StringType(), True), True
+            ),
+        ]
+    )
+    cleaned_ascwds_worker_schema = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(IndCQC.ascwds_worker_import_date, DateType(), True),
+            StructField(IndCQC.main_job_role_clean_labelled, StringType(), True),
+        ]
+    )
+
+
+class EstimateIndCQCFilledPostsByJobRoleUtilsSchemas:
+    test_map_column: str = "test_map_column"
+    create_map_column_schema = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(MainJobRoleLabels.care_worker, IntegerType(), True),
+            StructField(MainJobRoleLabels.registered_nurse, IntegerType(), True),
+            StructField(MainJobRoleLabels.senior_care_worker, IntegerType(), True),
+            StructField(MainJobRoleLabels.senior_management, IntegerType(), True),
+        ]
+    )
+    expected_create_map_column_schema = StructType(
+        [
+            *create_map_column_schema,
+            StructField(
+                test_map_column,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+    expected_create_map_column_when_drop_columns_is_true_schema = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(
+                test_map_column,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+
+    aggregate_ascwds_worker_with_additional_column_schema = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(IndCQC.ascwds_worker_import_date, DateType(), True),
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.main_job_role_clean_labelled, StringType(), True),
+        ]
+    )
+
+    aggregate_ascwds_worker_schema = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(IndCQC.ascwds_worker_import_date, DateType(), True),
+            StructField(IndCQC.main_job_role_clean_labelled, StringType(), True),
+        ]
+    )
+    expected_aggregate_ascwds_worker_schema = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(IndCQC.ascwds_worker_import_date, DateType(), True),
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+
+    estimated_filled_posts_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.care_home, StringType(), True),
+            StructField(IndCQC.number_of_beds, IntegerType(), True),
+            StructField(IndCQC.ascwds_workplace_import_date, DateType(), True),
+            StructField(IndCQC.establishment_id, StringType(), True),
+        ]
+    )
+    aggregated_job_role_breakdown_df = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(IndCQC.ascwds_worker_import_date, DateType(), True),
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+    merged_job_role_estimate_schema = StructType(
+        [
+            *estimated_filled_posts_schema,
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+
+    create_total_from_values_in_map_column_when_counts_are_longs_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), LongType()),
+                True,
+            ),
+        ]
+    )
+    expected_create_total_from_values_in_map_column_when_counts_are_longs_schema = (
+        StructType(
+            [
+                *create_total_from_values_in_map_column_when_counts_are_longs_schema,
+                StructField("temp_total_count_of_worker_records", LongType(), True),
+            ]
+        )
+    )
+
+    create_total_from_values_in_map_column_when_counts_are_doubles_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), DoubleType()),
+                True,
+            ),
+        ]
+    )
+    expected_create_total_from_values_in_map_column_when_counts_are_doubles_schema = (
+        StructType(
+            [
+                *create_total_from_values_in_map_column_when_counts_are_doubles_schema,
+                StructField("temp_total_count_of_worker_records", DoubleType(), True),
+            ]
+        )
+    )
+
+    create_ratios_from_counts_when_counts_are_longs_schema = (
+        expected_create_total_from_values_in_map_column_when_counts_are_longs_schema
+    )
+    expected_create_ratios_from_counts_when_counts_are_longs_schema = StructType(
+        [
+            *create_ratios_from_counts_when_counts_are_longs_schema,
+            StructField(
+                IndCQC.ascwds_job_role_ratios, MapType(StringType(), DoubleType()), True
+            ),
+        ]
+    )
+
+    create_ratios_from_counts_when_counts_are_doubles_schema = (
+        expected_create_total_from_values_in_map_column_when_counts_are_doubles_schema
+    )
+    expected_create_ratios_from_counts_when_counts_are_doubles_schema = StructType(
+        [
+            *create_ratios_from_counts_when_counts_are_doubles_schema,
+            StructField(
+                IndCQC.ascwds_job_role_ratios,
+                MapType(StringType(), DoubleType()),
+                True,
+            ),
+        ]
+    )
+
+    create_estimate_filled_posts_by_job_role_map_column_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.estimate_filled_posts, DoubleType(), True),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_merged,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_create_estimate_filled_posts_by_job_role_map_column_schema = StructType(
+        [
+            *create_estimate_filled_posts_by_job_role_map_column_schema,
+            StructField(
+                IndCQC.estimate_filled_posts_by_job_role,
+                MapType(StringType(), DoubleType()),
+                True,
+            ),
+        ]
+    )
+
+    remove_ascwds_job_role_count_when_estimate_filled_posts_source_not_ascwds_schema = (
+        StructType(
+            [
+                StructField(IndCQC.location_id, StringType(), True),
+                StructField(IndCQC.ascwds_filled_posts_dedup_clean, DoubleType(), True),
+                StructField(IndCQC.estimate_filled_posts, DoubleType(), True),
+                StructField(IndCQC.estimate_filled_posts_source, StringType(), True),
+                StructField(
+                    IndCQC.ascwds_job_role_counts,
+                    MapType(StringType(), IntegerType()),
+                    True,
+                ),
+            ]
+        )
+    )
+
+    count_registered_manager_names_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+            StructField(
+                IndCQC.registered_manager_names, ArrayType(StringType(), True), True
+            ),
+        ]
+    )
+    expected_count_registered_manager_names_schema = StructType(
+        [
+            *count_registered_manager_names_schema,
+            StructField(IndCQC.registered_manager_count, IntegerType(), True),
+        ]
+    )
+
+    sum_job_role_split_by_service_schema = StructType(
+        [
+            StructField(IndCQC.establishment_id, StringType(), True),
+            StructField(IndCQC.ascwds_worker_import_date, DateType(), True),
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+            StructField(IndCQC.primary_service_type, StringType(), True),
+        ]
+    )
+    expected_sum_job_role_split_by_service_schema = StructType(
+        [
+            *sum_job_role_split_by_service_schema,
+            StructField(
+                IndCQC.ascwds_job_role_rolling_sum,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+
+    interpolate_job_role_ratios_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_filtered,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_interpolate_job_role_ratios_schema = StructType(
+        [
+            *interpolate_job_role_ratios_schema,
+            StructField(
+                IndCQC.ascwds_job_role_ratios_interpolated,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+
+    extrapolate_job_role_ratios_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_filtered,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_extrapolate_job_role_ratios_schema = StructType(
+        [
+            *extrapolate_job_role_ratios_schema.fields,
+            StructField(
+                IndCQC.ascwds_job_role_ratios_extrapolated,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+
+    pivot_job_role_column_schema = StructType(
+        [
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(IndCQC.primary_service_type, StringType(), False),
+            StructField(IndCQC.main_job_role_clean_labelled, StringType(), False),
+            StructField(IndCQC.ascwds_job_role_ratios_interpolated, FloatType(), True),
+        ]
+    )
+    expected_pivot_job_role_column_schema = StructType(
+        [
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(IndCQC.primary_service_type, StringType(), False),
+            StructField(MainJobRoleLabels.care_worker, FloatType(), True),
+            StructField(MainJobRoleLabels.registered_nurse, FloatType(), True),
+            StructField(MainJobRoleLabels.senior_care_worker, FloatType(), True),
+            StructField(MainJobRoleLabels.senior_management, FloatType(), True),
+        ]
+    )
+    expected_pivot_job_role_column_two_job_roles_schema = StructType(
+        [
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(IndCQC.primary_service_type, StringType(), False),
+            StructField(MainJobRoleLabels.care_worker, FloatType(), True),
+            StructField(MainJobRoleLabels.registered_nurse, FloatType(), True),
+        ]
+    )
+
+    convert_map_with_all_null_values_to_null_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.unix_time, StringType(), False),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_interpolated,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_convert_map_with_all_null_values_to_null_schema = StructType(
+        [
+            *convert_map_with_all_null_values_to_null_schema,
+        ]
+    )
+
+    unpacked_mapped_column_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+            StructField(
+                IndCQC.estimate_filled_posts_by_job_role,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_unpacked_mapped_column_schema = StructType(
+        [
+            *unpacked_mapped_column_schema,
+            StructField(MainJobRoleLabels.care_worker, FloatType(), True),
+            StructField(MainJobRoleLabels.registered_nurse, FloatType(), True),
+            StructField(MainJobRoleLabels.senior_care_worker, FloatType(), True),
+            StructField(MainJobRoleLabels.senior_management, FloatType(), True),
+        ]
+    )
+
+    non_rm_managerial_estimate_filled_posts_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(MainJobRoleLabels.senior_management, FloatType(), True),
+            StructField(MainJobRoleLabels.middle_management, FloatType(), True),
+            StructField(MainJobRoleLabels.first_line_manager, FloatType(), True),
+            StructField(MainJobRoleLabels.supervisor, FloatType(), True),
+            StructField(MainJobRoleLabels.other_managerial_staff, FloatType(), True),
+            StructField(MainJobRoleLabels.deputy_manager, FloatType(), True),
+            StructField(MainJobRoleLabels.team_leader, FloatType(), True),
+            StructField(MainJobRoleLabels.data_governance_manager, FloatType(), True),
+            StructField(MainJobRoleLabels.it_manager, FloatType(), True),
+            StructField(MainJobRoleLabels.it_service_desk_manager, FloatType(), True),
+        ]
+    )
+    expected_non_rm_managerial_estimate_filled_posts_schema = StructType(
+        [
+            *non_rm_managerial_estimate_filled_posts_schema,
+            StructField(
+                IndCQC.sum_non_rm_managerial_estimated_filled_posts, FloatType(), True
+            ),
+            StructField(
+                IndCQC.proportion_of_non_rm_managerial_estimated_filled_posts_by_role,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+
+    estimate_and_cqc_registered_manager_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(MainJobRoleLabels.registered_manager, FloatType(), True),
+            StructField(IndCQC.registered_manager_count, IntegerType(), True),
+        ]
+    )
+    expected_estimate_and_cqc_registered_manager_schema = StructType(
+        [
+            *estimate_and_cqc_registered_manager_schema,
+            StructField(
+                IndCQC.difference_between_estimate_and_cqc_registered_managers,
+                FloatType(),
+                True,
+            ),
+        ]
+    )
+
+    sum_job_group_counts_from_job_role_count_map_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+    expected_sum_job_group_counts_from_job_role_count_map_schema = StructType(
+        [
+            *sum_job_group_counts_from_job_role_count_map_schema,
+            StructField(
+                IndCQC.ascwds_job_group_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+    sum_job_group_counts_from_job_role_count_map_for_patching_create_map_column_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(
+                IndCQC.ascwds_job_group_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+
+    filter_ascwds_job_role_map_when_dc_or_manregprof_1_or_more_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.worker_records_bounded, IntegerType(), True),
+            StructField(
+                IndCQC.ascwds_job_role_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+            StructField(
+                IndCQC.ascwds_job_group_counts,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+    expected_filter_ascwds_job_role_map_when_dc_or_manregprof_1_or_more_schema = (
+        StructType(
+            [
+                *filter_ascwds_job_role_map_when_dc_or_manregprof_1_or_more_schema,
+                StructField(
+                    IndCQC.ascwds_job_role_counts_filtered,
+                    MapType(StringType(), IntegerType()),
+                    True,
+                ),
+            ]
+        )
+    )
+
+    filter_ascwds_job_role_count_map_when_job_group_ratios_outside_percentile_boundaries_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.primary_service_type, StringType(), True),
+            StructField(
+                IndCQC.ascwds_job_group_ratios,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+            StructField(
+                IndCQC.ascwds_job_role_counts_filtered,
+                MapType(StringType(), IntegerType()),
+                True,
+            ),
+        ]
+    )
+
+    transform_imputed_job_role_ratios_to_counts_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.estimate_filled_posts, FloatType(), True),
+            StructField(
+                IndCQC.imputed_ascwds_job_role_ratios,
+                MapType(StringType(), FloatType()),
+            ),
+        ]
+    )
+    expected_transform_imputed_job_role_ratios_to_counts_schema = StructType(
+        [
+            *transform_imputed_job_role_ratios_to_counts_schema,
+            StructField(
+                IndCQC.imputed_ascwds_job_role_counts,
+                MapType(StringType(), FloatType()),
+            ),
+        ]
+    )
+
+    recalculate_managerial_filled_posts_non_rm_col_list = [
+        "managerial_role_1",
+        "managerial_role_2",
+        "managerial_role_3",
+        "managerial_role_4",
+    ]
+    recalculate_managerial_filled_posts_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField("managerial_role_1", FloatType(), True),
+            StructField("managerial_role_2", FloatType(), True),
+            StructField("managerial_role_3", FloatType(), True),
+            StructField("managerial_role_4", FloatType(), True),
+            StructField(
+                IndCQC.proportion_of_non_rm_managerial_estimated_filled_posts_by_role,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+            StructField(
+                IndCQC.difference_between_estimate_and_cqc_registered_managers,
+                FloatType(),
+                True,
+            ),
+        ]
+    )
+
+    recalculate_total_filled_posts_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(MainJobRoleLabels.care_worker, FloatType(), False),
+            StructField(MainJobRoleLabels.registered_nurse, FloatType(), False),
+            StructField(MainJobRoleLabels.senior_care_worker, FloatType(), False),
+            StructField(MainJobRoleLabels.senior_management, FloatType(), False),
+        ]
+    )
+    expected_recalculate_total_filled_posts_schema = StructType(
+        [
+            *recalculate_total_filled_posts_schema,
+            StructField(
+                IndCQC.estimate_filled_posts_from_all_job_roles, FloatType(), False
+            ),
+        ]
+    )
+
+    overwrite_registered_manager_estimate_with_cqc_count_schema = StructType(
+        [
+            StructField(MainJobRoleLabels.registered_manager, FloatType(), False),
+            StructField(IndCQC.registered_manager_count, IntegerType(), False),
+        ]
+    )
+
+    combine_interpolated_and_extrapolated_job_role_ratios_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_filtered,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_interpolated,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+            StructField(
+                IndCQC.ascwds_job_role_ratios_extrapolated,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_combine_interpolated_and_extrapolated_job_role_ratios_schema = StructType(
+        [
+            *combine_interpolated_and_extrapolated_job_role_ratios_schema,
+            StructField(
+                IndCQC.imputed_ascwds_job_role_ratios,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+
+    calculate_difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles_schema = StructType(
+        [
+            StructField(IndCQC.estimate_filled_posts, FloatType(), False),
+            StructField(
+                IndCQC.estimate_filled_posts_from_all_job_roles, FloatType(), False
+            ),
+        ]
+    )
+    expected_calculate_difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles_schema = StructType(
+        [
+            *calculate_difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles_schema,
+            StructField(
+                IndCQC.difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles,
+                FloatType(),
+                False,
+            ),
+        ]
+    )
+
+
+@dataclass
+class EstimateJobRolesPrimaryServiceRollingSumSchemas:
+    add_rolling_sum_partitioned_by_primary_service_type_and_main_job_role_clean_labelled_schema = StructType(
+        [
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(IndCQC.primary_service_type, StringType(), False),
+            StructField(IndCQC.main_job_role_clean_labelled, StringType(), False),
+            StructField(IndCQC.ascwds_job_role_counts_exploded, FloatType(), True),
+        ]
+    )
+    expected_add_rolling_sum_partitioned_by_primary_service_type_and_main_job_role_clean_labelled_schema = StructType(
+        [
+            *add_rolling_sum_partitioned_by_primary_service_type_and_main_job_role_clean_labelled_schema,
+            StructField(IndCQC.ascwds_job_role_rolling_sum, FloatType(), True),
+        ]
+    )
+
+    primary_service_rolling_sum_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(IndCQC.primary_service_type, StringType(), False),
+            StructField(
+                IndCQC.imputed_ascwds_job_role_counts,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+    expected_primary_service_rolling_sum_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), False),
+            StructField(IndCQC.unix_time, IntegerType(), False),
+            StructField(IndCQC.primary_service_type, StringType(), False),
+            StructField(
+                IndCQC.imputed_ascwds_job_role_counts,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+            StructField(
+                IndCQC.ascwds_job_role_rolling_sum,
+                MapType(StringType(), FloatType()),
+                True,
+            ),
+        ]
+    )
+
+
+@dataclass
+class ValidateEstimatedIndCqcFilledPostsByJobRoleSchemas:
+    cleaned_ind_cqc_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+        ]
+    )
+    estimated_ind_cqc_filled_posts_by_job_role_schema = StructType(
+        [
+            StructField(IndCQC.location_id, StringType(), True),
+            StructField(IndCQC.cqc_location_import_date, DateType(), True),
+        ]
+    )
+    calculate_expected_size_schema = cleaned_ind_cqc_schema
