+++ tests/test_file_data.py
@@ -21,7 +21,6 @@ from utils.column_values.categorical_column_values import (
     Dormancy,
     EstimateFilledPostsSource,
     LocationType,
-    JobGroupLabels,
     MainJobRoleLabels,
     PrimaryServiceType,
     Region,
@@ -4714,32 +4713,6 @@ class ValidateEstimatedIndCqcFilledPostsData:
 
 
 @dataclass
-class ValidateEstimatedIndCqcFilledPostsByJobRoleData:
-    # fmt: off
-    cleaned_ind_cqc_rows = [
-        ("1-000000001", date(2024, 1, 1),),
-        ("1-000000002", date(2024, 1, 1),),
-        ("1-000000001", date(2024, 2, 1),),
-        ("1-000000002", date(2024, 2, 1),),
-    ]
-
-    estimated_ind_cqc_filled_posts_by_job_role_rows = [
-        ("1-000000001", date(2024, 1, 1),),
-        ("1-000000002", date(2024, 1, 1),),
-        ("1-000000001", date(2024, 1, 9),),
-        ("1-000000002", date(2024, 1, 9),),
-    ]
-    # fmt: on
-
-    calculate_expected_size_rows = [
-        (
-            "1-000000001",
-            date(2024, 1, 1),
-        ),
-    ]
-
-
-@dataclass
 class RawDataAdjustments:
     expected_worker_data = [
         ("worker_1", "20240101", "estab_1", "other"),
@@ -5530,2601 +5503,5 @@ class ValidateCleanedCapacityTrackerNonResData:
             "2024",
             "01",
             "01",
-        ),
-    ]
-
-
-@dataclass
-class EstimateIndCQCFilledPostsByJobRoleData:
-    estimated_ind_cqc_filled_posts_rows = [
-        (
-            "1-001",
-            date(2024, 1, 1),
-            "Service A",
-            "101",
-            date(2024, 1, 1),
-            3.0,
-            ["John Doe"],
-        ),
-        (
-            "1-002",
-            date(2025, 1, 1),
-            "Service A",
-            "101",
-            date(2025, 1, 1),
-            3.0,
-            ["John Doe"],
-        ),
-        (
-            "1-003",
-            date(2025, 1, 1),
-            "Service B",
-            "103",
-            date(2025, 1, 1),
-            3.0,
-            ["John Doe"],
-        ),
-        (
-            "1-004",
-            date(2025, 1, 1),
-            "Service A",
-            "104",
-            date(2025, 1, 1),
-            3.0,
-            ["John Doe"],
-        ),
-    ]
-    cleaned_ascwds_worker_rows = [
-        ("101", date(2024, 1, 1), MainJobRoleLabels.senior_management),
-        ("101", date(2024, 1, 1), MainJobRoleLabels.care_worker),
-        ("101", date(2024, 1, 1), MainJobRoleLabels.care_worker),
-        ("101", date(2025, 1, 1), MainJobRoleLabels.care_worker),
-        ("101", date(2025, 1, 1), MainJobRoleLabels.care_worker),
-        ("103", date(2025, 1, 1), MainJobRoleLabels.senior_management),
-        ("103", date(2025, 1, 1), MainJobRoleLabels.registered_nurse),
-        ("103", date(2025, 1, 1), MainJobRoleLabels.care_worker),
-        ("103", date(2025, 1, 1), MainJobRoleLabels.care_worker),
-        ("103", date(2025, 1, 1), MainJobRoleLabels.care_worker),
-        ("103", date(2025, 1, 1), MainJobRoleLabels.care_worker),
-        ("111", date(2025, 1, 1), MainJobRoleLabels.care_worker),
-    ]
-
-
-@dataclass
-class EstimateIndCQCFilledPostsByJobRoleUtilsData:
-    list_of_job_roles_for_tests = [
-        MainJobRoleLabels.care_worker,
-        MainJobRoleLabels.registered_nurse,
-        MainJobRoleLabels.senior_care_worker,
-        MainJobRoleLabels.senior_management,
-    ]
-
-    aggregate_ascwds_worker_job_roles_per_establishment_rows = [
-        ("101", date(2024, 1, 1), "1-001", MainJobRoleLabels.care_worker),
-    ]
-
-    aggregate_ascwds_worker_job_roles_per_establishment_when_all_job_roles_present_rows = [
-        ("101", date(2024, 1, 1), MainJobRoleLabels.care_worker),
-        ("101", date(2024, 1, 1), MainJobRoleLabels.care_worker),
-        ("101", date(2024, 1, 1), MainJobRoleLabels.registered_nurse),
-        ("102", date(2024, 1, 1), MainJobRoleLabels.senior_care_worker),
-        ("102", date(2024, 1, 1), MainJobRoleLabels.senior_management),
-        ("102", date(2024, 1, 2), MainJobRoleLabels.care_worker),
-    ]
-    expected_aggregate_ascwds_worker_job_roles_per_establishment_when_all_job_roles_present_rows = [
-        (
-            "101",
-            date(2024, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 2,
-                MainJobRoleLabels.registered_nurse: 1,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.senior_management: 0,
-            },
-        ),
-        (
-            "102",
-            date(2024, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 0,
-                MainJobRoleLabels.senior_care_worker: 1,
-                MainJobRoleLabels.senior_management: 1,
-            },
-        ),
-        (
-            "102",
-            date(2024, 1, 2),
-            {
-                MainJobRoleLabels.care_worker: 1,
-                MainJobRoleLabels.registered_nurse: 0,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.senior_management: 0,
-            },
-        ),
-    ]
-
-    aggregate_ascwds_worker_job_roles_per_establishment_when_some_job_roles_never_present_rows = [
-        ("101", date(2024, 1, 1), MainJobRoleLabels.senior_management),
-        ("101", date(2024, 1, 1), MainJobRoleLabels.registered_nurse),
-    ]
-    expected_aggregate_ascwds_worker_job_roles_per_establishment_when_some_job_roles_never_present_rows = [
-        (
-            "101",
-            date(2024, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 1,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.senior_management: 1,
-            },
-        ),
-    ]
-
-    aggregate_ascwds_worker_job_roles_per_establishment_missing_roles_replaced_with_zero_rows = [
-        ("101", date(2024, 1, 1), MainJobRoleLabels.registered_nurse),
-    ]
-    expected_aggregate_ascwds_worker_job_roles_per_establishment_missing_roles_replaced_with_zero_rows = [
-        (
-            "101",
-            date(2024, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 1,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.senior_management: 0,
-            },
-        ),
-    ]
-
-    aggregate_ascwds_worker_job_roles_per_establishment_when_single_establishment_has_multiple_dates_rows = [
-        ("101", date(2024, 1, 1), MainJobRoleLabels.senior_care_worker),
-        ("101", date(2024, 1, 1), MainJobRoleLabels.senior_management),
-        ("101", date(2024, 1, 2), MainJobRoleLabels.care_worker),
-    ]
-    expected_aggregate_ascwds_worker_job_roles_per_establishment_when_single_establishment_has_multiple_dates_rows = [
-        (
-            "101",
-            date(2024, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 0,
-                MainJobRoleLabels.senior_care_worker: 1,
-                MainJobRoleLabels.senior_management: 1,
-            },
-        ),
-        (
-            "101",
-            date(2024, 1, 2),
-            {
-                MainJobRoleLabels.care_worker: 1,
-                MainJobRoleLabels.registered_nurse: 0,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.senior_management: 0,
-            },
-        ),
-    ]
-
-    aggregate_ascwds_worker_job_roles_per_establishment_when_multiple_establishments_on_the_same_date_rows = [
-        ("101", date(2024, 1, 1), MainJobRoleLabels.senior_care_worker),
-        ("101", date(2024, 1, 1), MainJobRoleLabels.senior_management),
-        ("102", date(2024, 1, 1), MainJobRoleLabels.care_worker),
-    ]
-    expected_aggregate_ascwds_worker_job_roles_per_establishment_when_multiple_establishments_on_the_same_date_rows = [
-        (
-            "101",
-            date(2024, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 0,
-                MainJobRoleLabels.senior_care_worker: 1,
-                MainJobRoleLabels.senior_management: 1,
-            },
-        ),
-        (
-            "101",
-            date(2024, 1, 2),
-            {
-                MainJobRoleLabels.care_worker: 1,
-                MainJobRoleLabels.registered_nurse: 0,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.senior_management: 0,
-            },
-        ),
-    ]
-
-    aggregate_ascwds_worker_job_roles_per_establishment_when_unrecognised_role_present_rows = [
-        ("101", date(2024, 1, 1), MainJobRoleLabels.senior_care_worker),
-        ("101", date(2024, 1, 1), "unrecognised_role"),
-    ]
-    expected_aggregate_ascwds_worker_job_roles_per_establishment_when_unrecognised_role_present_rows = [
-        (
-            "101",
-            date(2024, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 0,
-                MainJobRoleLabels.senior_care_worker: 1,
-                MainJobRoleLabels.senior_management: 0,
-            },
-        ),
-    ]
-
-    create_map_column_when_all_columns_populated_rows = [("123", 0, 10, 20, 30)]
-    expected_create_map_column_when_all_columns_populated_rows = [
-        (
-            "123",
-            0,
-            10,
-            20,
-            30,
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 10,
-                MainJobRoleLabels.senior_care_worker: 20,
-                MainJobRoleLabels.senior_management: 30,
-            },
-        )
-    ]
-    expected_create_map_column_when_all_columns_populated_and_drop_columns_is_true_rows = [
-        (
-            "123",
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: 10,
-                MainJobRoleLabels.senior_care_worker: 20,
-                MainJobRoleLabels.senior_management: 30,
-            },
-        )
-    ]
-    create_map_column_when_some_columns_populated_rows = [("123", 0, None, 20, None)]
-    expected_create_map_column_when_some_columns_populated_rows = [
-        (
-            "123",
-            0,
-            None,
-            20,
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: 20,
-                MainJobRoleLabels.senior_management: None,
-            },
-        )
-    ]
-    create_map_column_when_no_columns_populated_rows = [("123", None, None, None, None)]
-    expected_create_map_column_when_no_columns_populated_rows = [
-        (
-            "123",
-            None,
-            None,
-            None,
-            None,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-        )
-    ]
-
-    # fmt: off
-    estimated_filled_posts_when_single_establishment_has_multiple_dates_rows = [
-        ("1-1001", CareHome.care_home, 3, date(2025, 1, 1), "1"),
-        ("1-1001", CareHome.care_home, 5, date(2025, 2, 1), "1"),
-        ("1-1001", CareHome.care_home, 7, date(2025, 3, 1), "1"),
-    ]
-    aggregated_job_role_breakdown_when_single_establishment_has_multiple_dates_rows = [
-        ("1", date(2025, 1, 1), {MainJobRoleLabels.care_worker: 0, MainJobRoleLabels.registered_nurse: 1}),
-        ("1", date(2025, 3, 1), {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1", date(2025, 5, 1), {MainJobRoleLabels.care_worker: 2, MainJobRoleLabels.registered_nurse: 3}),
-    ]
-    expected_merge_dataframse_when_single_establishment_has_multiple_dates_rows = [
-        ("1-1001", CareHome.care_home, 3, date(2025, 1, 1), "1", {MainJobRoleLabels.care_worker: 0, MainJobRoleLabels.registered_nurse: 1}),
-        ("1-1001", CareHome.care_home, 5, date(2025, 2, 1), "1", None),
-        ("1-1001", CareHome.care_home, 7, date(2025, 3, 1), "1", {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-    ]
-    # fmt: on
-
-    # fmt: off
-    estimated_filled_posts_when_multiple_establishments_on_the_same_date_rows = [
-        ("1-1001", CareHome.care_home, 3, date(2025, 1, 1), "1"),
-        ("1-1002", CareHome.care_home, 5, date(2025, 1, 1), "2"),
-        ("1-1003", CareHome.care_home, 7, date(2025, 1, 1), "3"),
-    ]
-    aggregated_job_role_breakdown_when_multiple_establishments_on_the_same_date_rows = [
-        ("1", date(2025, 1, 1), {MainJobRoleLabels.care_worker: 0, MainJobRoleLabels.registered_nurse: 1}),
-        ("2", date(2025, 1, 1), {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-    ]
-    expected_merge_dataframse_when_multiple_establishments_on_the_same_date_rows = [
-        ("1-1001", CareHome.care_home, 3, date(2025, 1, 1), "1", {MainJobRoleLabels.care_worker: 0, MainJobRoleLabels.registered_nurse: 1}),
-        ("1-1002", CareHome.care_home, 5, date(2025, 1, 1), "2", {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1-1003", CareHome.care_home, 7, date(2025, 1, 1), "3", None),
-    ]
-    # fmt: on
-
-    # fmt: off
-    estimated_filled_posts_when_establishments_do_not_match_rows = [
-        ("1-1001", CareHome.care_home, 3, date(2025, 1, 1), "1"),
-    ]
-    aggregated_job_role_breakdown_when_establishments_do_not_match_rows = [
-        ("2", date(2025, 1, 1), {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-    ]
-    expected_merge_dataframse_when_establishments_do_not_match_rows = [
-        ("1-1001", CareHome.care_home, 3, date(2025, 1, 1), "1", None),
-    ]
-    # fmt: on
-
-    temp_total_count_of_worker_records = "temp_total_count_of_worker_records"
-
-    create_total_from_values_in_map_column_when_counts_are_longs_rows = [
-        (
-            "1-001",
-            {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2},
-        ),
-        (
-            "1-002",
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-            },
-        ),
-        (
-            "1-003",
-            None,
-        ),
-    ]
-    expected_create_total_from_values_in_map_column_when_counts_are_longs_rows = [
-        (
-            "1-001",
-            {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2},
-            3,
-        ),
-        (
-            "1-002",
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-            },
-            None,
-        ),
-        (
-            "1-003",
-            None,
-            None,
-        ),
-    ]
-
-    create_total_from_values_in_map_column_when_counts_are_doubles_rows = [
-        (
-            "1-001",
-            {
-                MainJobRoleLabels.care_worker: 2.0,
-                MainJobRoleLabels.registered_nurse: 3.0,
-            },
-        ),
-    ]
-    expected_create_total_from_values_in_map_column_when_counts_are_doubles_rows = [
-        (
-            "1-001",
-            {
-                MainJobRoleLabels.care_worker: 2.0,
-                MainJobRoleLabels.registered_nurse: 3.0,
-            },
-            5.0,
-        ),
-    ]
-
-    create_ratios_from_counts_when_counts_are_longs_rows = (
-        expected_create_total_from_values_in_map_column_when_counts_are_longs_rows
-    )
-    expected_create_ratios_from_counts_when_counts_are_longs_rows = [
-        (
-            "1-001",
-            {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2},
-            3,
-            {
-                MainJobRoleLabels.care_worker: 0.333,
-                MainJobRoleLabels.registered_nurse: 0.667,
-            },
-        ),
-        (
-            "1-002",
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-            },
-            None,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-            },
-        ),
-        (
-            "1-003",
-            None,
-            None,
-            None,
-        ),
-    ]
-
-    create_ratios_from_counts_when_counts_are_doubles_rows = (
-        expected_create_total_from_values_in_map_column_when_counts_are_doubles_rows
-    )
-    expected_create_ratios_from_counts_when_counts_are_doubles_rows = [
-        (
-            "1-001",
-            {
-                MainJobRoleLabels.care_worker: 2.0,
-                MainJobRoleLabels.registered_nurse: 3.0,
-            },
-            5.0,
-            {
-                MainJobRoleLabels.care_worker: 0.4,
-                MainJobRoleLabels.registered_nurse: 0.6,
-            },
-        ),
-    ]
-
-    # fmt: off
-    create_estimate_filled_posts_by_job_role_map_column_when_all_job_role_ratios_populated_rows = [
-        ("1-001",
-         100.0,
-        {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5})
-    ]
-
-    expected_create_estimate_filled_posts_by_job_role_map_column_when_all_job_role_ratios_populated_rows = [
-        ("1-001",
-         100.0,
-        {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-        {MainJobRoleLabels.care_worker: 50.0, MainJobRoleLabels.registered_nurse: 50.0})
-    ]
-    # fmt: on
-
-    create_estimate_filled_posts_by_job_role_map_column_when_job_role_ratio_column_is_null_rows = [
-        ("1-001", 100.0, None)
-    ]
-
-    expected_create_estimate_filled_posts_by_job_role_map_column_when_job_role_ratio_column_is_null_rows = [
-        ("1-001", 100.0, None, None)
-    ]
-
-    # fmt: off
-    create_estimate_filled_posts_by_job_role_map_column_when_estimate_filled_posts_is_null_rows = [
-        (
-            "1-001",
-            None,
-            {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-        )
-    ]
-
-    expected_create_estimate_filled_posts_by_job_role_map_column_when_estimate_filled_posts_is_null_rows = [
-        (
-            "1-001",
-            None,
-            {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-            {MainJobRoleLabels.care_worker: None, MainJobRoleLabels.registered_nurse: None},
-        )
-    ]
-    # fmt: on
-
-    # fmt: off
-    remove_ascwds_job_role_count_when_estimate_filled_posts_source_not_ascwds_rows = [
-        ("1-001",
-         10.0,
-         10.0,
-         EstimateFilledPostsSource.ascwds_pir_merged,
-         {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1-002",
-         None,
-         20.0,
-         EstimateFilledPostsSource.ascwds_pir_merged,
-         {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1-003",
-         10.0,
-         10.0,
-         EstimateFilledPostsSource.care_home_model,
-         {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-    ]
-    # fmt: on
-
-    # fmt: off
-    expected_remove_ascwds_job_role_count_when_estimate_filled_posts_source_not_ascwds_rows = [
-        ("1-001",
-         10.0,
-         10.0,
-         EstimateFilledPostsSource.ascwds_pir_merged,
-         {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1-002",
-         None,
-         20.0,
-         EstimateFilledPostsSource.ascwds_pir_merged,
-         None),
-        ("1-003",
-         10.0,
-         10.0,
-         EstimateFilledPostsSource.care_home_model,
-         None),
-    ]
-    # fmt: on
-
-    count_registered_manager_names_when_location_has_one_registered_manager_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe"])
-    ]
-    expected_count_registered_manager_names_when_location_has_one_registered_manager_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe"], 1)
-    ]
-
-    count_registered_manager_names_when_location_has_two_registered_managers_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe", "Jane Doe"])
-    ]
-    expected_count_registered_manager_names_when_location_has_two_registered_managers_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe", "Jane Doe"], 1)
-    ]
-
-    count_registered_manager_names_when_location_has_null_registered_manager_rows = [
-        ("1-0000000001", date(2025, 1, 1), None)
-    ]
-    expected_count_registered_manager_names_when_location_has_null_registered_manager_rows = [
-        ("1-0000000001", date(2025, 1, 1), None, 0)
-    ]
-
-    count_registered_manager_names_when_location_has_empty_list_rows = [
-        ("1-0000000001", date(2025, 1, 1), [])
-    ]
-    expected_count_registered_manager_names_when_location_has_empty_list_rows = [
-        ("1-0000000001", date(2025, 1, 1), [], 0)
-    ]
-
-    count_registered_manager_names_when_two_locations_have_different_number_of_registered_managers_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe"]),
-        ("1-0000000002", date(2025, 1, 1), ["John Doe", "Jane Doe"]),
-    ]
-    expected_count_registered_manager_names_when_two_locations_have_different_number_of_registered_managers_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe"], 1),
-        ("1-0000000002", date(2025, 1, 1), ["John Doe", "Jane Doe"], 1),
-    ]
-
-    count_registered_manager_names_when_a_location_has_different_number_of_registered_managers_at_different_import_dates_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe"]),
-        ("1-0000000001", date(2025, 2, 1), ["John Doe", "Jane Doe"]),
-    ]
-    expected_count_registered_manager_names_when_a_location_has_different_number_of_registered_managers_at_different_import_dates_rows = [
-        ("1-0000000001", date(2025, 1, 1), ["John Doe"], 1),
-        ("1-0000000001", date(2025, 2, 1), ["John Doe", "Jane Doe"], 1),
-    ]
-
-    unpacked_mapped_column_with_one_record_data = [
-        (
-            "1-001",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        )
-    ]
-    expected_unpacked_mapped_column_with_one_record_data = [
-        (
-            "1-001",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-            1.0,
-            2.0,
-            3.0,
-            4.0,
-        )
-    ]
-
-    unpacked_mapped_column_with_map_items_in_different_orders_data = [
-        (
-            "1-001",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 4.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-            },
-        ),
-        (
-            "1-002",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.senior_management: 40.0,
-                MainJobRoleLabels.registered_nurse: 20.0,
-                MainJobRoleLabels.care_worker: 10.0,
-                MainJobRoleLabels.senior_care_worker: 30.0,
-            },
-        ),
-    ]
-    expected_unpacked_mapped_column_with_map_items_in_different_orders_data = [
-        (
-            "1-001",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 4.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-            },
-            1.0,
-            2.0,
-            3.0,
-            4.0,
-        ),
-        (
-            "1-002",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.senior_management: 40.0,
-                MainJobRoleLabels.registered_nurse: 20.0,
-                MainJobRoleLabels.care_worker: 10.0,
-                MainJobRoleLabels.senior_care_worker: 30.0,
-            },
-            10.0,
-            20.0,
-            30.0,
-            40.0,
-        ),
-    ]
-
-    unpacked_mapped_column_with_null_values_data = [
-        (
-            "1-001",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-        ),
-        (
-            "1-002",
-            date(2025, 2, 1),
-            None,
-        ),
-    ]
-    expected_unpacked_mapped_column_with_null_values_data = [
-        (
-            "1-001",
-            date(2025, 1, 1),
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-            None,
-            None,
-            None,
-            None,
-        ),
-        (
-            "1-002",
-            date(2025, 2, 1),
-            None,
-            None,
-            None,
-            None,
-            None,
-        ),
-    ]
-
-    # fmt: off
-    non_rm_managerial_estimate_filled_posts_rows = [
-        ("1-001", 9.0, 3.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0),
-        ("1-002", 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0),
-    ]
-
-    expected_non_rm_managerial_estimate_filled_posts_rows = [
-        ("1-001", 9.0, 3.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0,
-         20.0,
-         {MainJobRoleLabels.senior_management: 0.45,
-          MainJobRoleLabels.middle_management: 0.15,
-          MainJobRoleLabels.first_line_manager: 0.05,
-          MainJobRoleLabels.supervisor: 0.05,
-          MainJobRoleLabels.other_managerial_staff: 0.05,
-          MainJobRoleLabels.deputy_manager: 0.05,
-          MainJobRoleLabels.team_leader: 0.05,
-          MainJobRoleLabels.data_governance_manager: 0.05,
-          MainJobRoleLabels.it_manager: 0.05,
-          MainJobRoleLabels.it_service_desk_manager: 0.05},
-         ),
-        ("1-002", 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0,
-         0.0,
-         {MainJobRoleLabels.senior_management: 0.1,
-          MainJobRoleLabels.middle_management: 0.1,
-          MainJobRoleLabels.first_line_manager: 0.1,
-          MainJobRoleLabels.supervisor: 0.1,
-          MainJobRoleLabels.other_managerial_staff: 0.1,
-          MainJobRoleLabels.deputy_manager: 0.1,
-          MainJobRoleLabels.team_leader: 0.1,
-          MainJobRoleLabels.data_governance_manager: 0.1,
-          MainJobRoleLabels.it_manager: 0.1,
-          MainJobRoleLabels.it_service_desk_manager: 0.1}
-         ),
-    ]
-
-    # fmt: on
-    interpolate_job_role_ratios_data = [
-        (
-            "1000",
-            1000,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 1.0,
-                MainJobRoleLabels.senior_care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 1.0,
-            },
-        ),
-        (
-            "1000",
-            1001,
-            None,
-        ),
-        (
-            "1000",
-            1002,
-            {
-                MainJobRoleLabels.care_worker: 3.0,
-                MainJobRoleLabels.registered_nurse: 3.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 3.0,
-            },
-        ),
-    ]
-
-    expected_interpolate_job_role_ratios_data = [
-        (
-            "1000",
-            1000,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 1.0,
-                MainJobRoleLabels.senior_care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 1.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 1.0,
-                MainJobRoleLabels.senior_care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 1.0,
-            },
-        ),
-        (
-            "1000",
-            1001,
-            None,
-            {
-                MainJobRoleLabels.care_worker: 2.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 2.0,
-                MainJobRoleLabels.senior_management: 2.0,
-            },
-        ),
-        (
-            "1000",
-            1002,
-            {
-                MainJobRoleLabels.care_worker: 3.0,
-                MainJobRoleLabels.registered_nurse: 3.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 3.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 3.0,
-                MainJobRoleLabels.registered_nurse: 3.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 3.0,
-            },
-        ),
-    ]
-
-    interpolate_job_role_ratios_with_null_records_which_cannot_be_interpolated_data = [
-        (
-            "1000",
-            1000,
-            None,
-        ),
-        (
-            "1000",
-            1001,
-            {
-                MainJobRoleLabels.care_worker: 2.0,
-                MainJobRoleLabels.registered_nurse: 4.0,
-                MainJobRoleLabels.senior_care_worker: 6.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-        ),
-        (
-            "1000",
-            1002,
-            None,
-        ),
-        (
-            "1000",
-            1003,
-            {
-                MainJobRoleLabels.care_worker: 0.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 4.0,
-                MainJobRoleLabels.senior_management: 5.0,
-            },
-        ),
-        (
-            "1000",
-            1004,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 1.0,
-                MainJobRoleLabels.senior_care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 1.0,
-            },
-        ),
-        (
-            "1000",
-            1005,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 4.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 2.0,
-            },
-        ),
-        (
-            "1000",
-            1006,
-            None,
-        ),
-    ]
-
-    expected_interpolate_job_role_ratios_with_null_records_which_cannot_be_interpolated_data = [
-        (
-            "1000",
-            1000,
-            None,
-            None,
-        ),
-        (
-            "1000",
-            1001,
-            {
-                MainJobRoleLabels.care_worker: 2.0,
-                MainJobRoleLabels.registered_nurse: 4.0,
-                MainJobRoleLabels.senior_care_worker: 6.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 2.0,
-                MainJobRoleLabels.registered_nurse: 4.0,
-                MainJobRoleLabels.senior_care_worker: 6.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-        ),
-        (
-            "1000",
-            1002,
-            None,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 3.0,
-                MainJobRoleLabels.senior_care_worker: 5.0,
-                MainJobRoleLabels.senior_management: 6.5,
-            },
-        ),
-        (
-            "1000",
-            1003,
-            {
-                MainJobRoleLabels.care_worker: 0.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 4.0,
-                MainJobRoleLabels.senior_management: 5.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 4.0,
-                MainJobRoleLabels.senior_management: 5.0,
-            },
-        ),
-        (
-            "1000",
-            1004,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 1.0,
-                MainJobRoleLabels.senior_care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 1.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 1.0,
-                MainJobRoleLabels.senior_care_worker: 1.0,
-                MainJobRoleLabels.senior_management: 1.0,
-            },
-        ),
-        (
-            "1000",
-            1005,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 4.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 2.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 4.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 2.0,
-            },
-        ),
-        (
-            "1000",
-            1006,
-            None,
-            None,
-        ),
-    ]
-
-    # fmt: off
-    pivot_job_role_column_returns_expected_pivot_rows = [
-        (1000, PrimaryServiceType.care_home_only, MainJobRoleLabels.care_worker, 1.0),
-        (1000, PrimaryServiceType.care_home_only, MainJobRoleLabels.registered_nurse, 2.0),
-        (1000, PrimaryServiceType.care_home_only, MainJobRoleLabels.senior_care_worker, 3.0),
-        (1000, PrimaryServiceType.care_home_only, MainJobRoleLabels.senior_management, 4.0),
-    ]
-    expected_pivot_job_role_column_returns_expected_pivot_rows = [
-        (1000,PrimaryServiceType.care_home_only, 1.0, 2.0, 3.0, 4.0),
-    ]
-    # fmt: on
-
-    # fmt: off
-    pivot_job_role_column_with_multiple_grouping_column_options_rows = [
-        (1000, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.care_worker, 1.0),
-        (1000, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.registered_nurse, 6.0),
-        (1001, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.care_worker, 2.0),
-        (1001, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.registered_nurse, 5.0),
-        (1000, PrimaryServiceType.care_home_only, MainJobRoleLabels.care_worker, 3.0),
-        (1000, PrimaryServiceType.care_home_only, MainJobRoleLabels.registered_nurse, 4.0),
-        (1001, PrimaryServiceType.care_home_only, MainJobRoleLabels.care_worker, 4.0),
-        (1001, PrimaryServiceType.care_home_only, MainJobRoleLabels.registered_nurse, 3.0),
-        (1000, PrimaryServiceType.non_residential, MainJobRoleLabels.care_worker, 5.0),
-        (1000, PrimaryServiceType.non_residential, MainJobRoleLabels.registered_nurse, 2.0),
-        (1001, PrimaryServiceType.non_residential, MainJobRoleLabels.care_worker, 6.0),
-        (1001, PrimaryServiceType.non_residential, MainJobRoleLabels.registered_nurse, 1.0),
-        (1002, PrimaryServiceType.non_residential, MainJobRoleLabels.care_worker, None),
-        (1002, PrimaryServiceType.non_residential, MainJobRoleLabels.registered_nurse, None),
-    ]
-    expected_pivot_job_role_column_with_multiple_grouping_column_options_rows = [
-        (1000, PrimaryServiceType.care_home_with_nursing, 1.0, 6.0),
-        (1000, PrimaryServiceType.care_home_only, 3.0, 4.0),
-        (1000, PrimaryServiceType.non_residential, 5.0, 2.0),
-        (1001, PrimaryServiceType.care_home_with_nursing, 2.0, 5.0),
-        (1001, PrimaryServiceType.care_home_only, 4.0, 3.0),
-        (1001, PrimaryServiceType.non_residential, 6.0, 1.0),
-        (1002, PrimaryServiceType.non_residential, None, None),
-    ]
-    # fmt: on
-
-    # fmt: off
-    pivot_job_role_column_returns_first_aggregation_column_value_rows = [
-        (1000, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.care_worker, 1.0),
-        (1000, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.care_worker, 2.0),
-        (1000, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.registered_nurse, 3.0),
-        (1000, PrimaryServiceType.care_home_with_nursing, MainJobRoleLabels.registered_nurse, 4.0),
-        (1001, PrimaryServiceType.care_home_only, MainJobRoleLabels.care_worker, 5.0),
-        (1001, PrimaryServiceType.care_home_only, MainJobRoleLabels.care_worker, 6.0),
-        (1002, PrimaryServiceType.care_home_only, MainJobRoleLabels.registered_nurse, 7.0),
-        (1002, PrimaryServiceType.care_home_only, MainJobRoleLabels.registered_nurse, 8.0),
-    ]
-    expected_pivot_job_role_column_returns_first_aggregation_column_value_rows = [
-        (1000, PrimaryServiceType.care_home_with_nursing, 1.0, 3.0),
-        (1001, PrimaryServiceType.care_home_only, 5.0, None),
-        (1002, PrimaryServiceType.care_home_only, None, 7.0),
-    ]
-    # fmt: on
-
-    convert_map_with_all_null_values_to_null_map_has_no_nulls_data = [
-        (
-            "1000",
-            1,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-        )
-    ]
-
-    expected_convert_map_with_all_null_values_to_null_map_has_no_nulls_data = [
-        (
-            "1000",
-            1,
-            None,
-        )
-    ]
-
-    convert_map_with_all_null_values_to_null_map_has_all_nulls = [
-        (
-            "1001",
-            1,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        )
-    ]
-
-    expected_convert_map_with_all_null_values_to_null_map_has_all_nulls = [
-        (
-            "1001",
-            1,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        )
-    ]
-
-    convert_map_with_all_null_values_to_null_when_map_has_all_null_and_all_non_null_records_data = [
-        (
-            "2000",
-            1,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-        ),
-        (
-            "2001",
-            1,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-        ),
-    ]
-
-    expected_convert_map_with_all_null_values_to_null_when_map_has_all_null_and_all_non_null_records_data = [
-        (
-            "2000",
-            1,
-            None,
-        ),
-        (
-            "2001",
-            1,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-        ),
-    ]
-
-    convert_map_with_all_null_values_to_null_when_map_has_some_nulls_data = [
-        (
-            "1002",
-            1,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: None,
-            },
-        )
-    ]
-
-    expected_convert_map_with_all_null_values_to_null_when_map_has_some_nulls_data = [
-        (
-            "1002",
-            1,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: None,
-            },
-        )
-    ]
-
-    estimate_and_cqc_registered_manager_rows = [
-        ("1-001", 0.0, 1),
-        ("1-002", 10.0, 1),
-        ("1-003", None, 1),
-        ("1-004", 10.0, None),
-    ]
-    expected_estimate_and_cqc_registered_manager_rows = [
-        ("1-001", 0.0, 1, -1.0),
-        ("1-002", 10.0, 1, 9.0),
-        ("1-003", None, 1, None),
-        ("1-004", 10.0, None, None),
-    ]
-
-    sum_job_group_counts_from_job_role_count_map_rows = [
-        (
-            "1-001",
-            1000,
-            {
-                MainJobRoleLabels.care_worker: 1,
-                MainJobRoleLabels.senior_care_worker: 1,
-                MainJobRoleLabels.senior_management: 2,
-                MainJobRoleLabels.first_line_manager: 2,
-                MainJobRoleLabels.registered_nurse: 3,
-                MainJobRoleLabels.social_worker: 3,
-                MainJobRoleLabels.admin_staff: 4,
-                MainJobRoleLabels.ancillary_staff: 4,
-            },
-        ),
-        (
-            "1-001",
-            1001,
-            {
-                MainJobRoleLabels.care_worker: 10,
-                MainJobRoleLabels.senior_care_worker: 10,
-                MainJobRoleLabels.senior_management: 20,
-                MainJobRoleLabels.first_line_manager: 20,
-                MainJobRoleLabels.registered_nurse: 30,
-                MainJobRoleLabels.social_worker: 30,
-                MainJobRoleLabels.admin_staff: 40,
-                MainJobRoleLabels.ancillary_staff: 40,
-            },
-        ),
-        (
-            "1-002",
-            1000,
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.registered_nurse: None,
-            },
-        ),
-        (
-            "1-003",
-            1000,
-            None,
-        ),
-    ]
-    expected_sum_job_group_counts_from_job_role_count_map_rows = [
-        (
-            "1-001",
-            1000,
-            {
-                MainJobRoleLabels.care_worker: 1,
-                MainJobRoleLabels.senior_care_worker: 1,
-                MainJobRoleLabels.senior_management: 2,
-                MainJobRoleLabels.first_line_manager: 2,
-                MainJobRoleLabels.registered_nurse: 3,
-                MainJobRoleLabels.social_worker: 3,
-                MainJobRoleLabels.admin_staff: 4,
-                MainJobRoleLabels.ancillary_staff: 4,
-            },
-            {
-                JobGroupLabels.direct_care: 2,
-                JobGroupLabels.managers: 4,
-                JobGroupLabels.regulated_professions: 6,
-                JobGroupLabels.other: 8,
-            },
-        ),
-        (
-            "1-001",
-            1001,
-            {
-                MainJobRoleLabels.care_worker: 10,
-                MainJobRoleLabels.senior_care_worker: 10,
-                MainJobRoleLabels.senior_management: 20,
-                MainJobRoleLabels.first_line_manager: 20,
-                MainJobRoleLabels.registered_nurse: 30,
-                MainJobRoleLabels.social_worker: 30,
-                MainJobRoleLabels.admin_staff: 40,
-                MainJobRoleLabels.ancillary_staff: 40,
-            },
-            {
-                JobGroupLabels.direct_care: 20,
-                JobGroupLabels.managers: 40,
-                JobGroupLabels.regulated_professions: 60,
-                JobGroupLabels.other: 80,
-            },
-        ),
-        (
-            "1-002",
-            1000,
-            {
-                MainJobRoleLabels.care_worker: 0,
-                MainJobRoleLabels.senior_care_worker: 0,
-                MainJobRoleLabels.registered_nurse: None,
-            },
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 0,
-            },
-        ),
-        (
-            "1-003",
-            1000,
-            None,
-            None,
-        ),
-    ]
-    sum_job_group_counts_from_job_role_count_map_for_patching_create_map_column_rows = [
-        (
-            "1-001",
-            1001,
-            {
-                JobGroupLabels.direct_care: 20,
-                JobGroupLabels.managers: 40,
-                JobGroupLabels.regulated_professions: 60,
-                JobGroupLabels.other: 80,
-            },
-        ),
-        (
-            "1-002",
-            1000,
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 0,
-            },
-        ),
-        (
-            "1-001",
-            1000,
-            {
-                JobGroupLabels.direct_care: 2,
-                JobGroupLabels.managers: 4,
-                JobGroupLabels.regulated_professions: 6,
-                JobGroupLabels.other: 8,
-            },
-        ),
-    ]
-
-    filter_ascwds_job_role_map_when_dc_or_manregprof_1_or_more_rows = [
-        (
-            "1-001",
-            None,
-            None,
-            None,
-        ),
-        (
-            "1-002",
-            0,
-            None,
-            None,
-        ),
-        (
-            "1-003",
-            1,
-            {
-                MainJobRoleLabels.admin_staff: 0,
-            },
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 1,
-            },
-        ),
-        (
-            "1-004",
-            1,
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-            {
-                JobGroupLabels.direct_care: 1,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 0,
-            },
-        ),
-        (
-            "1-005",
-            1,
-            {
-                MainJobRoleLabels.senior_management: 1,
-            },
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 1,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 0,
-            },
-        ),
-        (
-            "1-006",
-            1,
-            {
-                MainJobRoleLabels.social_worker: 1,
-            },
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 1,
-                JobGroupLabels.other: 0,
-            },
-        ),
-    ]
-    expected_filter_ascwds_job_role_map_when_dc_or_manregprof_1_or_more_rows = [
-        (
-            "1-001",
-            None,
-            None,
-            None,
-            None,
-        ),
-        (
-            "1-002",
-            0,
-            None,
-            None,
-            None,
-        ),
-        (
-            "1-003",
-            1,
-            {
-                MainJobRoleLabels.admin_staff: 0,
-            },
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 1,
-            },
-            None,
-        ),
-        (
-            "1-004",
-            1,
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-            {
-                JobGroupLabels.direct_care: 1,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-005",
-            1,
-            {
-                MainJobRoleLabels.senior_management: 1,
-            },
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 1,
-                JobGroupLabels.regulated_professions: 0,
-                JobGroupLabels.other: 0,
-            },
-            {
-                MainJobRoleLabels.senior_management: 1,
-            },
-        ),
-        (
-            "1-006",
-            1,
-            {
-                MainJobRoleLabels.social_worker: 1,
-            },
-            {
-                JobGroupLabels.direct_care: 0,
-                JobGroupLabels.managers: 0,
-                JobGroupLabels.regulated_professions: 1,
-                JobGroupLabels.other: 0,
-            },
-            {
-                MainJobRoleLabels.social_worker: 1,
-            },
-        ),
-    ]
-
-    filter_ascwds_job_role_count_map_when_job_group_ratios_outside_percentile_boundaries_rows = [
-        (
-            "1-001",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 10.0,
-                JobGroupLabels.managers: 0.9,
-                JobGroupLabels.regulated_professions: 0.9,
-                JobGroupLabels.other: 0.9,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-002",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.9,
-                JobGroupLabels.managers: 10.0,
-                JobGroupLabels.regulated_professions: 0.8,
-                JobGroupLabels.other: 0.8,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-003",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.8,
-                JobGroupLabels.managers: 0.8,
-                JobGroupLabels.regulated_professions: 10.0,
-                JobGroupLabels.other: 0.7,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-004",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.7,
-                JobGroupLabels.managers: 0.7,
-                JobGroupLabels.regulated_professions: 0.7,
-                JobGroupLabels.other: 10.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-005",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.0,
-                JobGroupLabels.managers: 0.6,
-                JobGroupLabels.regulated_professions: 0.6,
-                JobGroupLabels.other: 0.6,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-006",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.6,
-                JobGroupLabels.managers: 0.5,
-                JobGroupLabels.regulated_professions: 0.5,
-                JobGroupLabels.other: 0.5,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-007",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 10.0,
-                JobGroupLabels.managers: 0.9,
-                JobGroupLabels.regulated_professions: 0.9,
-                JobGroupLabels.other: 0.9,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-008",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.9,
-                JobGroupLabels.managers: 10.0,
-                JobGroupLabels.regulated_professions: 0.8,
-                JobGroupLabels.other: 0.8,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-009",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.8,
-                JobGroupLabels.managers: 0.8,
-                JobGroupLabels.regulated_professions: 10.0,
-                JobGroupLabels.other: 0.7,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-010",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.7,
-                JobGroupLabels.managers: 0.7,
-                JobGroupLabels.regulated_professions: 0.7,
-                JobGroupLabels.other: 10.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-011",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.0,
-                JobGroupLabels.managers: 0.6,
-                JobGroupLabels.regulated_professions: 0.6,
-                JobGroupLabels.other: 0.6,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-012",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.6,
-                JobGroupLabels.managers: 0.5,
-                JobGroupLabels.regulated_professions: 0.5,
-                JobGroupLabels.other: 0.5,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-    ]
-    expected_filter_ascwds_job_role_count_map_when_job_group_ratios_outside_percentile_boundaries_rows = [
-        (
-            "1-001",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 10.0,
-                JobGroupLabels.managers: 0.9,
-                JobGroupLabels.regulated_professions: 0.9,
-                JobGroupLabels.other: 0.9,
-            },
-            None,
-        ),
-        (
-            "1-002",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.9,
-                JobGroupLabels.managers: 10.0,
-                JobGroupLabels.regulated_professions: 0.8,
-                JobGroupLabels.other: 0.8,
-            },
-            None,
-        ),
-        (
-            "1-003",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.8,
-                JobGroupLabels.managers: 0.8,
-                JobGroupLabels.regulated_professions: 10.0,
-                JobGroupLabels.other: 0.7,
-            },
-            None,
-        ),
-        (
-            "1-004",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.7,
-                JobGroupLabels.managers: 0.7,
-                JobGroupLabels.regulated_professions: 0.7,
-                JobGroupLabels.other: 10.0,
-            },
-            None,
-        ),
-        (
-            "1-005",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.0,
-                JobGroupLabels.managers: 0.6,
-                JobGroupLabels.regulated_professions: 0.6,
-                JobGroupLabels.other: 0.6,
-            },
-            None,
-        ),
-        (
-            "1-006",
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                JobGroupLabels.direct_care: 0.6,
-                JobGroupLabels.managers: 0.5,
-                JobGroupLabels.regulated_professions: 0.5,
-                JobGroupLabels.other: 0.5,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-        (
-            "1-007",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 10.0,
-                JobGroupLabels.managers: 0.9,
-                JobGroupLabels.regulated_professions: 0.9,
-                JobGroupLabels.other: 0.9,
-            },
-            None,
-        ),
-        (
-            "1-008",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.9,
-                JobGroupLabels.managers: 10.0,
-                JobGroupLabels.regulated_professions: 0.8,
-                JobGroupLabels.other: 0.8,
-            },
-            None,
-        ),
-        (
-            "1-009",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.8,
-                JobGroupLabels.managers: 0.8,
-                JobGroupLabels.regulated_professions: 10.0,
-                JobGroupLabels.other: 0.7,
-            },
-            None,
-        ),
-        (
-            "1-010",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.7,
-                JobGroupLabels.managers: 0.7,
-                JobGroupLabels.regulated_professions: 0.7,
-                JobGroupLabels.other: 10.0,
-            },
-            None,
-        ),
-        (
-            "1-011",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.0,
-                JobGroupLabels.managers: 0.6,
-                JobGroupLabels.regulated_professions: 0.6,
-                JobGroupLabels.other: 0.6,
-            },
-            None,
-        ),
-        (
-            "1-012",
-            PrimaryServiceType.care_home_only,
-            {
-                JobGroupLabels.direct_care: 0.6,
-                JobGroupLabels.managers: 0.5,
-                JobGroupLabels.regulated_professions: 0.5,
-                JobGroupLabels.other: 0.5,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1,
-            },
-        ),
-    ]
-
-    transform_imputed_job_role_ratios_to_counts_rows = [
-        (
-            "1-001",
-            100.0,
-            {
-                MainJobRoleLabels.care_worker: 0.10,
-                MainJobRoleLabels.registered_nurse: 0.20,
-                MainJobRoleLabels.senior_care_worker: 0.30,
-                MainJobRoleLabels.senior_management: 0.40,
-            },
-        ),
-        (
-            "1-002",
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.10,
-                MainJobRoleLabels.registered_nurse: 0.20,
-                MainJobRoleLabels.senior_care_worker: 0.30,
-                MainJobRoleLabels.senior_management: 0.40,
-            },
-        ),
-        (
-            "1-003",
-            100.0,
-            None,
-        ),
-    ]
-    expected_transform_imputed_job_role_ratios_to_counts_rows = [
-        (
-            "1-001",
-            100.0,
-            {
-                MainJobRoleLabels.care_worker: 0.10,
-                MainJobRoleLabels.registered_nurse: 0.20,
-                MainJobRoleLabels.senior_care_worker: 0.30,
-                MainJobRoleLabels.senior_management: 0.40,
-            },
-            {
-                MainJobRoleLabels.care_worker: 10.0,
-                MainJobRoleLabels.registered_nurse: 20.0,
-                MainJobRoleLabels.senior_care_worker: 30.0,
-                MainJobRoleLabels.senior_management: 40.0,
-            },
-        ),
-        (
-            "1-002",
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.10,
-                MainJobRoleLabels.registered_nurse: 0.20,
-                MainJobRoleLabels.senior_care_worker: 0.30,
-                MainJobRoleLabels.senior_management: 0.40,
-            },
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-        ),
-        (
-            "1-003",
-            100.0,
-            None,
-            None,
-        ),
-    ]
-
-    # fmt: off
-    job_role_ratios_extrapolation_rows = [
-        (
-            "1-001",
-            1000000200,
-            None
-        ),
-        (
-            "1-001",
-            1000000300,
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-        ),
-        (
-            "1-001",
-            1000000400,
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            }
-        ),
-        (
-            "1-001",
-            1000000500,
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-        ),
-        (
-            "1-001",
-            1000000600,
-            None,
-        ),
-        (
-            "1-002",
-            1000000200,
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            }
-        ),
-        (
-            "1-002",
-            1000000300,
-            None
-        ),
-        (
-            "1-002",
-            1000000400,
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-        ),
-        ("1-003", 1000000200, None),
-        ("1-003", 1000000300, None),
-        ("1-003", 1000000400, None),
-        ("1-003", 1000000500, None),
-    ]
-    expected_job_role_ratios_extrapolation_rows = [
-        (
-            "1-001",
-            1000000200,
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-        ),
-        (
-            "1-001",
-            1000000300,
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-        ),
-        (
-            "1-001",
-            1000000400,
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-        ),
-        (
-            "1-001",
-            1000000500,
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-        ),
-        (
-            "1-001",
-            1000000600,
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-        ),
-        (
-            "1-002",
-            1000000200,
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-        ),
-        (
-            "1-002",
-            1000000300,
-            None,
-            None,
-        ),
-        (
-            "1-002",
-            1000000400,
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-        ),
-        ("1-003", 1000000200, None, None),
-        ("1-003", 1000000300, None, None),
-        ("1-003", 1000000400, None, None),
-        ("1-003", 1000000500, None, None),
-    ]
-    # fmt: on
-
-    recalculate_managerial_filled_posts_rows = [
-        (
-            "1-001",
-            0.0,
-            1.0,
-            4.0,
-            5.0,
-            {
-                "managerial_role_1": 0.0,
-                "managerial_role_2": 0.1,
-                "managerial_role_3": 0.4,
-                "managerial_role_4": 0.5,
-            },
-            1.0,
-        ),
-        (
-            "1-002",
-            1.0,
-            1.0,
-            1.0,
-            1.0,
-            {
-                "managerial_role_1": 0.25,
-                "managerial_role_2": 0.25,
-                "managerial_role_3": 0.25,
-                "managerial_role_4": 0.25,
-            },
-            0.0,
-        ),
-        (
-            "1-003",
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            {
-                "managerial_role_1": 0.25,
-                "managerial_role_2": 0.25,
-                "managerial_role_3": 0.25,
-                "managerial_role_4": 0.25,
-            },
-            -1.0,
-        ),
-        (
-            "1-004",
-            0.0,
-            1.0,
-            4.0,
-            5.0,
-            {
-                "managerial_role_1": 0.0,
-                "managerial_role_2": 0.1,
-                "managerial_role_3": 0.4,
-                "managerial_role_4": 0.5,
-            },
-            -1.0,
-        ),
-    ]
-    expected_recalculate_managerial_filled_posts_rows = [
-        (
-            "1-001",
-            0.0,
-            1.1,
-            4.4,
-            5.5,
-            {
-                "managerial_role_1": 0.0,
-                "managerial_role_2": 0.1,
-                "managerial_role_3": 0.4,
-                "managerial_role_4": 0.5,
-            },
-            1.0,
-        ),
-        (
-            "1-002",
-            1.0,
-            1.0,
-            1.0,
-            1.0,
-            {
-                "managerial_role_1": 0.25,
-                "managerial_role_2": 0.25,
-                "managerial_role_3": 0.25,
-                "managerial_role_4": 0.25,
-            },
-            0.0,
-        ),
-        (
-            "1-003",
-            0.0,
-            0.0,
-            0.0,
-            0.0,
-            {
-                "managerial_role_1": 0.25,
-                "managerial_role_2": 0.25,
-                "managerial_role_3": 0.25,
-                "managerial_role_4": 0.25,
-            },
-            -1.0,
-        ),
-        (
-            "1-004",
-            0.0,
-            0.9,
-            3.6,
-            4.5,
-            {
-                "managerial_role_1": 0.0,
-                "managerial_role_2": 0.1,
-                "managerial_role_3": 0.4,
-                "managerial_role_4": 0.5,
-            },
-            -1.0,
-        ),
-    ]
-
-    recalculate_total_filled_posts_rows = [
-        ("1-001", 0.0, 0.0, 0.0, 0.0),
-        ("1-002", 2.0, 1.0, 2.0, 1.0),
-    ]
-    expected_recalculate_total_filled_posts_rows = [
-        ("1-001", 0.0, 0.0, 0.0, 0.0, 0.0),
-        ("1-002", 2.0, 1.0, 2.0, 1.0, 6.0),
-    ]
-
-    combine_interpolated_and_extrapolated_job_role_ratios_rows = [
-        (
-            "1-001",
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-        ),
-        (
-            "1-002",
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-        ),
-        (
-            "1-003",
-            None,
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-        ),
-        (
-            "1-004",
-            None,
-            None,
-            None,
-        ),
-    ]
-    expected_combine_interpolated_and_extrapolated_job_role_ratios_rows = [
-        (
-            "1-001",
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.1,
-                MainJobRoleLabels.registered_nurse: 0.1,
-            },
-        ),
-        (
-            "1-002",
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.2,
-                MainJobRoleLabels.registered_nurse: 0.2,
-            },
-        ),
-        (
-            "1-003",
-            None,
-            None,
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-            {
-                MainJobRoleLabels.care_worker: 0.3,
-                MainJobRoleLabels.registered_nurse: 0.3,
-            },
-        ),
-        (
-            "1-004",
-            None,
-            None,
-            None,
-            None,
-        ),
-    ]
-
-    overwrite_registered_manager_estimate_with_cqc_count_rows = [
-        (10.0, 1),
-        (10.0, 0),
-    ]
-    expected_overwrite_registered_manager_estimate_with_cqc_count_rows = [
-        (1.0, 1),
-        (0.0, 0),
-    ]
-
-    calculate_difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles_rows = [
-        (10.0, 10.0),
-        (10.0, 9.0),
-        (9.0, 10.0),
-    ]
-    expected_calculate_difference_between_estimate_filled_posts_and_estimate_filled_posts_from_all_job_roles_rows = [
-        (10.0, 10.0, 0.0),
-        (10.0, 9.0, -1.0),
-        (9.0, 10.0, 1.0),
-    ]
-
-
-@dataclass
-class EstimateJobRolesPrimaryServiceRollingSumData:
-    list_of_job_roles_for_tests = [
-        MainJobRoleLabels.care_worker,
-        MainJobRoleLabels.registered_nurse,
-        MainJobRoleLabels.senior_care_worker,
-        MainJobRoleLabels.senior_management,
-    ]
-
-    add_rolling_sum_partitioned_by_primary_service_type_and_main_job_role_clean_labelled_data = [
-        (
-            0,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            1.0,
-        ),
-        (
-            86401,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            1.0,
-        ),
-        (
-            86402,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            1.0,
-        ),
-        (
-            86403,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            None,
-        ),
-        (
-            86404,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.senior_management,
-            1.0,
-        ),
-        (
-            0,
-            PrimaryServiceType.care_home_only,
-            MainJobRoleLabels.care_worker,
-            None,
-        ),
-        (
-            86401,
-            PrimaryServiceType.care_home_only,
-            MainJobRoleLabels.care_worker,
-            1.0,
-        ),
-        (
-            0,
-            PrimaryServiceType.non_residential,
-            MainJobRoleLabels.care_worker,
-            10.0,
-        ),
-        (
-            86400,
-            PrimaryServiceType.non_residential,
-            MainJobRoleLabels.care_worker,
-            2.0,
-        ),
-        (
-            86401,
-            PrimaryServiceType.non_residential,
-            MainJobRoleLabels.care_worker,
-            8.0,
-        ),
-    ]
-
-    expected_add_rolling_sum_partitioned_by_primary_service_type_and_main_job_role_clean_labelled_data = [
-        (
-            0,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            1.0,
-            1.0,
-        ),
-        (
-            86401,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            1.0,
-            1.0,
-        ),
-        (
-            86402,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            1.0,
-            2.0,
-        ),
-        (
-            86403,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.care_worker,
-            None,
-            2.0,
-        ),
-        (
-            86404,
-            PrimaryServiceType.care_home_with_nursing,
-            MainJobRoleLabels.senior_management,
-            1.0,
-            1.0,
-        ),
-        (
-            0,
-            PrimaryServiceType.care_home_only,
-            MainJobRoleLabels.care_worker,
-            None,
-            None,
-        ),
-        (
-            86401,
-            PrimaryServiceType.care_home_only,
-            MainJobRoleLabels.care_worker,
-            1.0,
-            1.0,
-        ),
-        (
-            0,
-            PrimaryServiceType.non_residential,
-            MainJobRoleLabels.care_worker,
-            10.0,
-            10.0,
-        ),
-        (
-            86400,
-            PrimaryServiceType.non_residential,
-            MainJobRoleLabels.care_worker,
-            2.0,
-            12.0,
-        ),
-        (
-            86401,
-            PrimaryServiceType.non_residential,
-            MainJobRoleLabels.care_worker,
-            8.0,
-            10.0,
-        ),
-    ]
-
-    primary_service_rolling_sum_when_one_primary_service_present_rows = [
-        (
-            "1000",
-            1,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        ),
-        (
-            "1000",
-            2,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-        ),
-        (
-            "1000",
-            3,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-        ),
-    ]
-    expected_primary_service_rolling_sum_when_one_primary_service_present_rows = [
-        (
-            "1000",
-            1,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        ),
-        (
-            "1000",
-            2,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: None,
-                MainJobRoleLabels.registered_nurse: None,
-                MainJobRoleLabels.senior_care_worker: None,
-                MainJobRoleLabels.senior_management: None,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        ),
-        (
-            "1000",
-            3,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 6.0,
-                MainJobRoleLabels.registered_nurse: 8.0,
-                MainJobRoleLabels.senior_care_worker: 10.0,
-                MainJobRoleLabels.senior_management: 12.0,
-            },
-        ),
-    ]
-
-    primary_service_rolling_sum_when_multiple_primary_services_present_rows = [
-        (
-            "1000",
-            1,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        ),
-        (
-            "1000",
-            2,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-        ),
-        (
-            "1000",
-            1,
-            PrimaryServiceType.care_home_only,
-            {
-                MainJobRoleLabels.care_worker: 11.0,
-                MainJobRoleLabels.registered_nurse: 12.0,
-                MainJobRoleLabels.senior_care_worker: 13.0,
-                MainJobRoleLabels.senior_management: 14.0,
-            },
-        ),
-        (
-            "1000",
-            2,
-            PrimaryServiceType.care_home_only,
-            {
-                MainJobRoleLabels.care_worker: 15.0,
-                MainJobRoleLabels.registered_nurse: 16.0,
-                MainJobRoleLabels.senior_care_worker: 17.0,
-                MainJobRoleLabels.senior_management: 18.0,
-            },
-        ),
-    ]
-    expected_primary_service_rolling_sum_when_multiple_primary_services_present_rows = [
-        (
-            "1000",
-            1,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        ),
-        (
-            "1000",
-            2,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 6.0,
-                MainJobRoleLabels.registered_nurse: 8.0,
-                MainJobRoleLabels.senior_care_worker: 10.0,
-                MainJobRoleLabels.senior_management: 12.0,
-            },
-        ),
-        (
-            "1000",
-            1,
-            PrimaryServiceType.care_home_only,
-            {
-                MainJobRoleLabels.care_worker: 11.0,
-                MainJobRoleLabels.registered_nurse: 12.0,
-                MainJobRoleLabels.senior_care_worker: 13.0,
-                MainJobRoleLabels.senior_management: 14.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 11.0,
-                MainJobRoleLabels.registered_nurse: 12.0,
-                MainJobRoleLabels.senior_care_worker: 13.0,
-                MainJobRoleLabels.senior_management: 14.0,
-            },
-        ),
-        (
-            "1000",
-            2,
-            PrimaryServiceType.care_home_only,
-            {
-                MainJobRoleLabels.care_worker: 15.0,
-                MainJobRoleLabels.registered_nurse: 16.0,
-                MainJobRoleLabels.senior_care_worker: 17.0,
-                MainJobRoleLabels.senior_management: 18.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 26.0,
-                MainJobRoleLabels.registered_nurse: 28.0,
-                MainJobRoleLabels.senior_care_worker: 30.0,
-                MainJobRoleLabels.senior_management: 32.0,
-            },
-        ),
-    ]
-
-    primary_service_rolling_sum_when_days_not_within_rolling_window_rows = [
-        (
-            "1000",
-            1704067200,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        ),
-        (
-            "1000",
-            1720137600,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-        ),
-    ]
-    expected_primary_service_rolling_sum_when_days_not_within_rolling_window_rows = [
-        (
-            "1000",
-            1704067200,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 1.0,
-                MainJobRoleLabels.registered_nurse: 2.0,
-                MainJobRoleLabels.senior_care_worker: 3.0,
-                MainJobRoleLabels.senior_management: 4.0,
-            },
-        ),
-        (
-            "1000",
-            1720137600,
-            PrimaryServiceType.care_home_with_nursing,
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
-            {
-                MainJobRoleLabels.care_worker: 5.0,
-                MainJobRoleLabels.registered_nurse: 6.0,
-                MainJobRoleLabels.senior_care_worker: 7.0,
-                MainJobRoleLabels.senior_management: 8.0,
-            },
         ),
     ]
